Create 10_Sum_of_Subset_Problem.c
Write a program to find a subset of a given set S = {S 1 , S 2 ,. ... ,S n } of npositive integers whose sum
is equal to a given positive integer d. For example, if S ={1, 2,5, 6, 8} and d= 9, there are two
solutions {1,2,6}and {1,8}. Display a suitable message, ifthe given problem instance doesn't have
a solution.